% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{assemble_brickman}
\alias{assemble_brickman}
\title{Read, warp, subset and archive all combinations of the raw Brickman data}
\usage{
assemble_brickman(
  scenario = c("RCP85", "RCP45", "PRESENT"),
  year = c("2055", "2075", "PRESENT"),
  vars = "all",
  interval = c("mon", "ann"),
  bb = colby_bbox(),
  band_as_time = FALSE,
  path = brickman_path()
)
}
\arguments{
\item{scenario}{chr, one o rmore of "RCP85", "RCP45", or "PRESENT"}

\item{year}{chr, year "2055" or "2075"}

\item{vars}{chr, variable names (such as "SST" or "all")}

\item{interval}{chr, either "ann" or "mon" (default)}

\item{bb}{sf, something which provides a bounding box for subsetting}

\item{band_as_time}{logical, convert band to appropriate time/date stamp}

\item{path}{filepath, path to where you want data saved}
}
\value{
tibble database
}
\description{
Read, warp, subset and archive all combinations of the raw Brickman data
}
