% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{best_hyperparams}
\alias{best_hyperparams}
\title{function to extract the hyperparameters from the highest ranking model}
\usage{
best_hyperparams(ranked_tbl)
}
\arguments{
\item{ranked_tbl}{a ranked table of models ranked based on the average mean of metrics}
}
\value{
tibble of model metrics to be used to finalize the final workflow
}
\description{
function to extract the hyperparameters from the highest ranking model
}
