% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_attrs.default}
\alias{bind_attrs.default}
\title{Bind one or more `stars` objects by attribute.}
\usage{
\method{bind_attrs}{default}(x, ...)
}
\arguments{
\item{x}{a stars object}

\item{...}{one or more `stars` objects to bind to `x` as additional attributes.  Any NULL elements are
silently removed first. Ignored if `x` is a list.}
}
\value{
`stars` objects
}
\description{
Bind one or more `stars` objects by attribute.
}
\seealso{
[stars issue 440](https://github.com/r-spatial/stars/issues/440#issuecomment-877627732)
}
