% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_attrs.list}
\alias{bind_attrs.list}
\title{Bind a list of \code{stars} objects by attribute.}
\usage{
\method{bind_attrs}{list}(x, ...)
}
\arguments{
\item{x}{list of `stars` objects}

\item{...}{ignored}
}
\value{
`stars` objects
}
\description{
Bind a list of \code{stars} objects by attribute.
}
\seealso{
[stars issue 440](https://github.com/r-spatial/stars/issues/440#issuecomment-877627732)
}
