% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_attrs}
\alias{bind_attrs}
\title{bind_attrs generic}
\usage{
bind_attrs(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\value{
a `stars` object
}
\description{
bind_attrs generic
}
