% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_bands.list}
\alias{bind_bands.list}
\title{Bind a list of \code{stars} objects by attribute.}
\usage{
\method{bind_bands}{list}(x, ..., along = list(month = month.abb))
}
\arguments{
\item{x}{list of `stars` objects}

\item{...}{ignored}

\item{along}{named list indicating the name of the time dimension and its values}
}
\value{
`stars` objects
}
\description{
Bind a list of \code{stars} objects by attribute.
}
\seealso{
[stars issue 440](https://github.com/r-spatial/stars/issues/440#issuecomment-877627732)
}
