% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{bind_bands}
\alias{bind_bands}
\title{bind_bands generic}
\usage{
bind_bands(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\value{
a `stars` object
}
\description{
bind_bands generic
}
