% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{brickman_build_database}
\alias{brickman_build_database}
\title{Build the Brickman database}
\usage{
brickman_build_database(path = brickman_path())
}
\arguments{
\item{path}{chr, the path to the data}
}
\value{
tabular database
}
\description{
Build the Brickman database
}
