% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{brickman_database}
\alias{brickman_database}
\title{Read the Brickman database}
\usage{
brickman_database(path = file.path(ROOT_DATA_PATH, "brickman"))
}
\arguments{
\item{path}{chr, the path to the data}
}
\value{
tabular database
}
\description{
Read the Brickman database
}
