% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{brickman_path}
\alias{brickman_path}
\title{Retrieve the path to the brickman data}
\usage{
brickman_path(root = ROOT_DATA_PATH)
}
\arguments{
\item{root}{chr, the root global data directory}
}
\value{
the path to the brickman data directory
}
\description{
Retrieve the path to the brickman data
}
