% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{brickman_variables}
\alias{brickman_variables}
\title{Retrieve a listing of all Brickman variables}
\usage{
brickman_variables(interval = c("mon", "ann", "all")[1])
}
\arguments{
\item{interval}{chr, one of "mon" (monthly), "ann" (annual) or "all"}
}
\value{
table
}
\description{
Retrieve a listing of all Brickman variables
}
