% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{browse_obis}
\alias{browse_obis}
\title{Browse the record for one unique id}
\usage{
browse_obis(
  id = "00040fa1-7acd-4731-bf1e-6dc16e30c7d4",
  base_url = "https://api.obis.org/v3/occurrence"
)
}
\arguments{
\item{id}{chr, a unique ID or an OBIS record (one row) If multiples are 
provided all but the first are ignored.}

\item{root_url}{chr, the root url}
}
\value{
NULL invisibly
}
\description{
Browse the record for one unique id
}
\examples{
\dontrun{
  obs = read_obis(scientificname = "Mola mola")
  one = obis |> slice(202)
  browse_obis(one)
}
}
