% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{coastline_path}
\alias{coastline_path}
\title{Retrieve the path to the coastline data}
\usage{
coastline_path(root = ROOT_DATA_PATH)
}
\arguments{
\item{root}{chr, the root global data directory}
}
\value{
the path to the coastline data directory
}
\description{
Retrieve the path to the coastline data
}
