% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{colby_bbox}
\alias{colby_bbox}
\title{Retrieve the region of study coordinates}
\usage{
colby_bbox(
  x = c(xmin = -74.9, xmax = -65, ymin = 38.8, ymax = 46),
  form = "polygon"
)
}
\arguments{
\item{x}{num, named vector defining the bounding box.  We provide
a default but you can override these values.}

\item{form}{chr, names the output form of data as "vector",
"bbox" (from sf package), or "polygon" (from sf package)}
}
\value{
bbox or polygon sf object
}
\description{
Retrieve the region of study coordinates
}
