% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColbyForecastingDocs-package.R
\docType{package}
\name{ColbyForecastingDocs}
\alias{ColbyForecastingDocs-package}
\alias{ColbyForecastingDocs}
\title{ColbyForecastingDocs: Documents for auxillary functions}
\description{
Documentation for Colby Forecasting Courseware
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/ColbyForecastingDocs}
}

}
