% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{data_path}
\alias{data_path}
\title{Get the data path}
\usage{
data_path(..., root = personal_data_path())
}
\arguments{
\item{root}{chr, the root personal data directory}

\item{what}{chr path and file segments to append to the root path using the
system file separator}
}
\value{
some directory in the personal data path
}
\description{
Get the data path
}
