% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{depth_first}
\alias{depth_first}
\title{A function to reorder a stars object or data table so that depth comes first}
\usage{
depth_first(x)
}
\arguments{
\item{x}{stars object or a table of data}
}
\value{
the input possibly reordered if it has "depth" as a variable
}
\description{
A function to reorder a stars object or data table so that depth comes first
}
