% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{extract_brickman}
\alias{extract_brickman}
\title{Extract point data from a Brickman stars object}
\usage{
extract_brickman(x = read_brickman(), y = gom_buoys(), form = "long")
}
\arguments{
\item{x}{stars object (with or without month dimension)}

\item{y}{sf point data, if this has a "month" attribute then we extract by month
as long as x has a month dimension}

\item{form}{chr one of either "long" (the default) or "wide" to control 
output column layout.}

\item{...}{other arguments passed to `stars::st_extract()`}
}
\value{
table of variable data for each input point
}
\description{
Extract point data from a Brickman stars object
}
