% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{fetch_obis}
\alias{fetch_obis}
\title{Fetch one or more species of obis data}
\usage{
fetch_obis(
  scientificname = "Mola mola",
  bb = colby_bbox(),
  fields = standard_obis_fields(),
  tidy = TRUE,
  output_path = make_path(data_path("obis")),
  crs = 4326
)
}
\arguments{
\item{scientificname}{one or more "genus" or "genus species" scientific name(s)}

\item{bb}{either NULL (to skip) or something from which a bounding box may be
extracted.  Used to filter the request to a smaller-than-global scope.
The output oc `colby_bbox()` will do nicely.}

\item{fields}{chr one or more fields to retrieve}

\item{tidy}{logical, if TRUE tidy up the downloaded data}

\item{output_path}{chr, the destination path where the data are stored using 
Files are automatically named "genus_species.gpkg" or "genus.gpkg" as
appropriate for each request.}

\item{crs}{the CRS for the data}
}
\value{
spatial table of observations
}
\description{
Fetch one or more species of obis data
}
