% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{get_model_type}
\alias{get_model_type}
\title{A convenience to quickly get the type name of a model}
\usage{
get_model_type(x)
}
\arguments{
\item{x}{workflow object}
}
\value{
string value of the model type
}
\description{
A convenience to quickly get the type name of a model
}
