% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{get_response_type}
\alias{get_response_type}
\title{A little function to accept a model name and return the
appropriate value for the `type` argument to predict}
\usage{
get_response_type(name)
}
\arguments{
\item{name}{str, the model specification name}
}
\value{
str, the appropriate value for response type
}
\description{
A little function to accept a model name and return the
appropriate value for the `type` argument to predict
}
