% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_user}
\alias{get_user}
\title{A function to retrieve the username}
\usage{
get_user(strip_basename = TRUE)
}
\arguments{
\item{strip_basename}{logical, if TRUE return just the basename. 
This only has meaning if the platform is Windows, harmless
otherwise}
}
\value{
username
}
\description{
A function to retrieve the username
}
