% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_year}
\alias{get_year}
\title{A function to retrieve a dates year as a 4-digit character string}
\usage{
get_year(date = Sys.Date())
}
\arguments{
\item{date}{Date or POSIXct class object}
}
\value{
the year as 4 character string
}
\description{
A function to retrieve a dates year as a 4-digit character string
}
