% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buoys.R
\name{gom_buoys}
\alias{gom_buoys}
\title{Read in a table of buoy names and locations}
\usage{
gom_buoys(form = c("table", "sf")[2])
}
\arguments{
\item{form}{chr, one of 'table' or 'sf'}
}
\value{
either a table or sf table
}
\description{
Read in a table of buoy names and locations
}
