% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_covariates.R
\name{include_covariate}
\alias{include_covariate}
\title{Include one or more covariates to a list}
\usage{
include_covariate(x, add_me = "depth")
}
\arguments{
\item{x}{str a vector of covariates}

\item{add_me}{str a vector of one or more covariates to append to the input x}
}
\value{
a possibly updated version of the input x
}
\description{
Include one or more covariates to a list
}
