% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_docs.R
\name{install_docs}
\alias{install_docs}
\title{Install the Colby Forecasting Documentation package}
\usage{
install_docs(repos = "BigelowLab/ColbyForecastingDocs", ...)
}
\arguments{
\item{repos}{chr, the namespace and name of the package on github}

\item{...}{other arguments for \code{\link[remotes]{install_github}}}
}
\value{
see \code{\link[remotes]{install_github}}
}
\description{
Please see \href{https://github.com/BigelowLab/ColbyForecastingDocs}{ColbyForecastingDocs}
}
