% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{keep_under}
\alias{keep_under}
\title{Coerce a set of whole numbers between 1 and `under`.  Wrapping occurs such 
month 13 is associated with January while month 0 is associated with
December.}
\usage{
keep_under(x, under = 12)
}
\arguments{
\item{x}{one of more numbers, which will be rounded}

\item{under}{num, the upper range of numbers to coerce within}
}
\description{
Coerce a set of whole numbers between 1 and `under`.  Wrapping occurs such 
month 13 is associated with January while month 0 is associated with
December.
}
