% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{make_path}
\alias{make_path}
\title{Make a path if it doesn't exist}
\usage{
make_path(path)
}
\arguments{
\item{path}{chr the path to create}
}
\value{
the path
}
\description{
Make a path if it doesn't exist
}
