% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_table}
\alias{metrics_table}
\title{Function that takes a workflow set and returns a metric table,
possibly selecting hyperparameter set with the best average metric.}
\usage{
metrics_table(
  wf_set,
  wflow_id = dplyr::pull(wf_set, dplyr::all_of("wflow_id")),
  what = "best"
)
}
\arguments{
\item{wf_set}{a workflow set}

\item{wflow_id}{str, one or more workflow ids, by deafult all are computed}

\item{what}{str, currently accepts only "best" to chose the best set, otherwise
it returns all of the hyperparameters unranked.}
}
\value{
a table of models metrics, possible ranked by the average mean of metrics
}
\description{
Function that takes a workflow set and returns a metric table,
possibly selecting hyperparameter set with the best average metric.
}
