% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_confmat}
\alias{model_fit_confmat}
\title{Given a table of model fits, plot the confusion matrices}
\usage{
model_fit_confmat(
  x = read_model_fit(),
  wids = dplyr::pull(x, dplyr::all_of("wflow_id"))
)
}
\arguments{
\item{x}{table of model fits}

\item{wids}{str, the workflow(s) to operate upon, by default all of them}
}
\value{
faceted ggplot object
}
\description{
Borrows heavily from the [yardstick package](https://github.com/tidymodels/yardstick/blob/51761c4e7a34e960949c75aeb2952ef02c408106/R/conf_mat.R#L421)
}
