% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_pdp}
\alias{model_fit_pdp}
\title{Create a partial dependence plot for a model fit}
\usage{
model_fit_pdp(x, wid = dplyr::pull(x, dplyr::all_of("wflow_id"))[1], ...)
}
\arguments{
\item{x}{table of model fits}

\item{wid}{str the id of the workflow to plot (wflow_id)}

\item{...}{other arguments for the [partial_dependence_plot] function}
}
\value{
ggplot object
}
\description{
Create a partial dependence plot for a model fit
}
