% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_roc_auc}
\alias{model_fit_roc_auc}
\title{Given a table of model fits, plot the ROC curves with the AUC shown}
\usage{
model_fit_roc_auc(
  x = read_model_fit(),
  wids = dplyr::pull(x, dplyr::all_of("wflow_id"))
)
}
\arguments{
\item{x}{table of model fits}

\item{wids}{str, the workflow(s) to operate upon, by default all of them}
}
\value{
faceted ggplot object
}
\description{
Given a table of model fits, plot the ROC curves with the AUC shown
}
