% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_spec}
\alias{model_fit_spec}
\title{Retrieve the model specifications for one or more model fits}
\usage{
model_fit_spec(x)
}
\arguments{
\item{x}{workflow set table}
}
\value{
str vector of first element of class attribute of each workflowset
}
\description{
Retrieve the model specifications for one or more model fits
}
