% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_varimp}
\alias{model_fit_varimp}
\title{Compute permutation importance for model covariates}
\usage{
model_fit_varimp(
  x,
  wids = dplyr::pull(x, dplyr::all_of("wflow_id")),
  n = 5,
  arrange = c("none", "increasing", "decreasing")[1],
  ...
)
}
\arguments{
\item{x}{table of model fits}

\item{wids}{str, the workflow(s) to operate upon, by default all of them}

\item{n}{num, number of iterations}

\item{arrange}{char, one of "none" (default), "decreasing" or "increasing" to arrange
the output order}

\item{...}{other arguments for \code{\link[maxnet]{predict}} such
as \code{clamp} and \code{type}}
}
\value{
table of premutation importance scores (and raw values)
}
\description{
This is adapted from Peter D Wilson's [fitMaxnet R package](https://github.com/peterbat1/fitMaxnet),
as done by [K Oliveira](https://github.com/kolive4/carcharodon/blob/cedeca9d1812fddadfec3aa1fee7cf0195e43553/functions/tidysdm_fxns.R#L470).
}
