% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_varimp_plot}
\alias{model_fit_varimp_plot}
\title{Plot permutation importance for model covariates}
\usage{
model_fit_varimp_plot(
  x,
  wids = dplyr::pull(x, dplyr::all_of("wflow_id")),
  what = "mean",
  ...
)
}
\arguments{
\item{x}{table of model fits}

\item{wids}{str, the workflow(s) to operate upon, by default all of them}

\item{what}{str, the name of the metric to use}
}
\value{
ggplot
}
\description{
Plot permutation importance for model covariates
}
