% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_varnames}
\alias{model_fit_varnames}
\title{Retrieve the predictior variable names for a workflow}
\usage{
model_fit_varnames(x, wid = dplyr::pull(x, dplyr::all_of("wflow_id"))[1])
}
\arguments{
\item{x}{table of model fits}

\item{wid}{str the id of the workflow to plot (wflow_id)}
}
\value{
string vector of predictor names
}
\description{
Retrieve the predictior variable names for a workflow
}
