% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_vi}
\alias{model_fit_vi}
\title{Compute variable importance}
\usage{
model_fit_vi(
  x,
  wids = dplyr::pull(x, dplyr::all_of("wflow_id")),
  scale = TRUE,
  ...
)
}
\arguments{
\item{x}{table of model fits}

\item{wids}{str the id of the workflow to plot (wflow_id)}

\item{scale}{logical, if TRUE scale values 0-1 by model fit}

\item{...}{other arguments for the [vip::vi] function}
}
\value{
a table if variable importance values
}
\description{
Compute variable importance
}
