% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{model_fit_vip}
\alias{model_fit_vip}
\title{Make a heat map showing relative variable importance}
\usage{
model_fit_vip(x, ...)
}
\arguments{
\item{x}{table of model fits}

\item{...}{other arguments for `model_fit_vi`}
}
\value{
ggplot heat map
}
\description{
If glm is part of the group then POS and NEG are 
treated separately.
}
