% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{month_as_name}
\alias{month_as_name}
\title{Convert month numbers to month names (abbreviated)}
\usage{
month_as_name(x = seq_len(12), lut = month_lut())
}
\arguments{
\item{x}{num one or more month numbers}

\item{lut}{named month look up table}
}
\value{
month names (abbreviated)
}
\description{
Convert month numbers to month names (abbreviated)
}
