% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{month_as_number}
\alias{month_as_number}
\title{Convert month names  (abbreviated)  to month numbers}
\usage{
month_as_number(x = month.abb, lut = month_lut())
}
\arguments{
\item{x}{str one or more month names}

\item{lut}{named month look up table}
}
\value{
month numbers (1-12)
}
\description{
Convert month names  (abbreviated)  to month numbers
}
