% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month.R
\name{month_lut}
\alias{month_lut}
\title{Retrieve a look-up-table where the values are month numbers
and the names are month.abb}
\usage{
month_lut()
}
\value{
named vector of month numbers
}
\description{
Retrieve a look-up-table where the values are month numbers
and the names are month.abb
}
