% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{obis_path}
\alias{obis_path}
\title{Retrieve the path to the obis data}
\usage{
obis_path(root = personal_data_path())
}
\arguments{
\item{root}{chr, the root personal data directory}
}
\value{
the path to the brickman config file
}
\description{
Retrieve the path to the obis data
}
