% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence_plot}
\alias{partial_dependence_plot}
\title{Given a workflow or fitted model generate a partial dependence plot}
\usage{
partial_dependence_plot(
  x,
  share_y = "all",
  filename = NULL,
  v = NULL,
  data = NULL,
  outcome = "class",
  ...
)
}
\arguments{
\item{x}{workflow with a fitted model}

\item{share_y}{chr, by default "all" but see `?plot.EffectsData`}

\item{v}{chr, variable names `?plot.EffectsData`. Not needed for workflow.}

\item{data}{table with outcomes and predictors. Not needed for workflow.}

\item{outcome}{chr, the name of the column (variable) that defines the outcome}

\item{...}{other arguments for `?plot.EffectsData`}
}
\value{
plot object as ggplot or plotly see `?plot.EffectsData`
}
\description{
Given a workflow or fitted model generate a partial dependence plot
}
