% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{personal_data_path}
\alias{personal_data_path}
\title{Get the personal data path}
\usage{
personal_data_path(user = get_user(), year = get_year(), root = ROOT_DATA_PATH)
}
\arguments{
\item{user}{chr, the user's name (no spaces please)}

\item{year}{chr, the current year as 4-character string}

\item{root}{chr, the root directory for data}
}
\value{
chr personal data path
}
\description{
Get the personal data path
}
