% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot a single stars attribute (variable) over many months}
\usage{
plot_prediction(
  x,
  colors = c("magma", "inferno", "plasma", "viridis", "cividis", "rocket", "mako",
    "turbo")[1],
  coast = read_coastline(),
  coast_color = "white"
)
}
\arguments{
\item{x}{stars object with one or more months along the 'month' dimension}

\item{colors}{str the name of the color table to use.  See 
details here https://ggplot2-book.org/scales-colour.html#sec-colour-continuous.
This is ignored is the data type is not numeric.}

\item{coast}{sf coast line or NULL to skip}
}
\value{
a ggplot2 object
}
\description{
Plot a single stars attribute (variable) over many months
}
