% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot an annotated ROC with AUC}
\usage{
plot_roc(x, truth, pred, title = "ROC")
}
\arguments{
\item{x}{table of predictive outcomes - see `predict_model`}

\item{truth}{the truth column, usually `class`}

\item{pred}{the prediction column, usually .pred_presense}

\item{title}{chr the optional title}
}
\value{
ggplot2 object suitable for printing
}
\description{
Plot an annotated ROC with AUC
}
