% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_stars}
\alias{predict_stars}
\title{Predict a classification from stars object}
\usage{
predict_stars(wflow, newdata, type = "prob", threshold = 0.5, ...)
}
\arguments{
\item{wflow}{a workflow object}

\item{newdata}{stars data}

\item{threshold}{num the aribitray threshold used to define the outcome}

\item{...}{other arguments for `predict.model_fit`}

\item{typ}{chr the type of prediction, by default "prob" - see `?predict.model_fit`}
}
\value{
a stars object
}
\description{
Predict a classification from stars object
}
