% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_table}
\alias{predict_table}
\title{Predict a classification from a table of new data}
\usage{
predict_table(wflow, newdata, type = "prob", threshold = 0.5, ...)
}
\arguments{
\item{wflow}{a workflow object}

\item{newdata}{table of new data including a `class` variable}

\item{threshold}{num the aribitray threshold used to define the outcome}

\item{...}{other arguments for `predict.model_fit`}

\item{typ}{chr the type of prediction, by default "prob" - see `?predict.model_fit`}
}
\value{
a table of predictions with the orginal class label
}
\description{
Predict a classification from a table of new data
}
