% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast.R
\name{prep_coastlines}
\alias{prep_coastlines}
\title{Prepare the to "coarse", "medium" and "detail" coastlines}
\usage{
prep_coastlines(output_path = coastline_path(), bb = colby_bbox())
}
\arguments{
\item{output_path}{chr the path where we save the data}

\item{bb}{bounding box for clipping}
}
\value{
chr the names of the three saved files
}
\description{
Prepare the to "coarse", "medium" and "detail" coastlines
}
