% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_model_data.R
\name{prep_model_data}
\alias{prep_model_data}
\title{Given a selection of POINT model_input data, prepare the table needed to 
model}
\usage{
prep_model_data(
  model_input,
  month = "Jan",
  covars = NULL,
  depth = NULL,
  keep_vars = c(names(covars), "depth"),
  form = c("sf", "table")[1]
)
}
\arguments{
\item{model_input}{`sf` with at least one attribute ("class").}

\item{month}{the name of the month}

\item{covars}{NULL or stars object (with 12 months)}

\item{depth}{NULL or stars object with depth}

\item{keep_vars}{chr, a vector of variable to keep, if NULL keep all}

\item{form}{chr one of "sf" or "table" (to drop the spatial info)}
}
\value{
a plain table or `sf` table with attributes (variables) added
}
\description{
Given a selection of POINT model_input data, prepare the table needed to 
model
}
