% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_density.R
\name{rasterize_point_density}
\alias{rasterize_point_density}
\title{Given a set of points and a target raster geometry, compute the density (count)
of points in each cell.}
\usage{
rasterize_point_density(
  x,
  y,
  name = "count",
  dilate = 0,
  dilate_value = 1,
  mask = y
)
}
\arguments{
\item{x}{sf POINT object}

\item{y}{stars object that defines the geometry of the output}

\item{name}{chr, the name of the output variable}

\item{dilate}{num, the size of the square structuring element used for dilating the
output counts (padding with zeroes).  Set to 0 to skip.}

\item{dilate_value}{num, if dilation occurs, this is the value assigned to the padded cells}

\item{mask}{`stars` object that defines masked areas where
dilation does not occur with the value NA.  It must have the same spatial 
geometry as the input \code{y}.}
}
\value{
stars array with point densities
}
\description{
Given a set of points and a target raster geometry, compute the density (count)
of points in each cell.
}
