% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coast.R
\name{read_coastline}
\alias{read_coastline}
\title{Read a coastline geometry set}
\usage{
read_coastline(what = "medium", path = coastline_path())
}
\arguments{
\item{what}{chr one of "coarse", "medium" or "detail"}

\item{path}{chr the data path o the coastline products}
}
\value{
sfc coastline object
}
\description{
Read a coastline geometry set
}
