% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{read_configuration}
\alias{read_configuration}
\title{Given a version identifier, read a configuration}
\usage{
read_configuration(
  scientificname = "Canis familiaris",
  version = "v001",
  path = data_path("models")
)
}
\arguments{
\item{scientificname}{str the scientific name}

\item{version}{chr the version identifier (defaults to 'v001')}

\item{path}{chr you personal model data path (defaults to `data_path("models")`)}
}
\value{
configuration list
}
\description{
Given a version identifier, read a configuration
}
