% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{read_configurations}
\alias{read_configurations}
\title{Read one or more configurations into a table.}
\usage{
read_configurations(versions, path = data_path("models"))
}
\arguments{
\item{versions}{chr one or more versions}

\item{path}{the model data path}
}
\value{
table of config values
}
\description{
Note that if configurations differ in the number of elements, the output
will contains ALL elements (the union of elements) with missing items assigned
to NA or NULL as appropriate.
}
