% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{read_ensemble}
\alias{read_ensemble}
\title{Read an ensemble given the version and path}
\usage{
read_ensemble(version = "g008", path = data_path("models"))
}
\arguments{
\item{version}{chr the version to read}

\item{path}{chr the path to the data directory}
}
\value{
ensemble
}
\description{
Read an ensemble given the version and path
}
