% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{read_model}
\alias{read_model}
\title{Read a model given the version and path}
\usage{
read_model(version = "g_Aug", path = data_path("models"))
}
\arguments{
\item{version}{chr the version to read}

\item{path}{chr the path to the data directory}
}
\value{
model
}
\description{
Read a model given the version and path
}
