% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{read_model_fit}
\alias{read_model_fit}
\title{Read a workflowset fit}
\usage{
read_model_fit(
  filename = "model_fits",
  path = data_path("models"),
  packed = FALSE
)
}
\arguments{
\item{filename}{str a name for the file}

\item{path}{str or NULL, if not NULL then write each fitted workflow to file}

\item{packed}{logical, if TRUE unbundle the workflows}
}
\value{
a named list with one or more "last_fit" model objects
}
\description{
Read a workflowset fit
}
