% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{read_obis}
\alias{read_obis}
\title{Read one or more obis data files}
\usage{
read_obis(scientificname = "Mola mola", path = data_path("obis"))
}
\arguments{
\item{scientificname}{one or more scientific names to read}

\item{path}{chr the data path}
}
\value{
sf table
}
\description{
Read one or more obis data files
}
