% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{read_observations}
\alias{read_observations}
\title{Read raw OBIS data and then filter it}
\usage{
read_observations(scientificname = "Mola mola", minimum_year = 1970, ...)
}
\arguments{
\item{scientificname}{chr, the name of the species to read}

\item{minimum_year}{num, the earliest year of observation to accept or 
set to NULL to skip}

\item{...}{other arguments passed to `read_obis()`}
}
\value{
a filtered table of observations
}
\description{
Read raw OBIS data and then filter it
}
