% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{read_prediction}
\alias{read_prediction}
\title{Read a prediction from file}
\usage{
read_prediction(
  x,
  scientificname = "Mola mola",
  version = "v1",
  year = "CURRENT",
  scenario = "CURRENT",
  path = data_path("predictions")
)
}
\arguments{
\item{x}{stars object with one or more time layers}

\item{scientificname}{chr, the name of the species}

\item{version}{chr, the version identifier}

\item{year}{chr, one of CURRENT, 2055 or 2075}

\item{scenario}{chr, one of CURRENT, RCP45 or RCP85}

\item{path}{chr the data path}
}
\value{
the input object x (suitable for piping)
}
\description{
Read a prediction from file
}
