% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{read_sunspots}
\alias{read_sunspots}
\title{Read in a sunspot data}
\usage{
read_sunspots()
}
\value{
table of monthly sunspot activity
}
\description{
Read in a sunspot data
}
\examples{
read_sunspots() |>
 group_by(month) |>
 group_modify(
  function(tbl, key){
    dplyr::tibble(
    month = key$month,
    min = min(tbl$sunspots),
    q1 = quantile(tbl$sunspots, 0.25),
    mean = mean(tbl$sunspots),
    median = mean(tbl$sunspots),
    q3 = quantile(tbl$sunspots, 0.75),
    max = max(tbl$sunspots))
  })
}
