% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{read_workflow}
\alias{read_workflow}
\title{Read a workflow given the version and path}
\usage{
read_workflow(
  version = "Aug",
  type = "rand_forest",
  path = data_path("models")
)
}
\arguments{
\item{version}{str the version to read}

\item{type}{str the model type name ala "rand_forest"}

\item{path}{str the path to the data directory}
}
\value{
workflow
}
\description{
Read a workflow given the version and path
}
