% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_qmd}
\alias{scrape_qmd}
\title{Scrape the contents of a .qmd to a .R file}
\usage{
scrape_qmd(
  src = "C00_coding.qmd",
  dst = file.path(get_username(), gsub(".qmd", ".R", basename(src), fixed = TRUE))
)
}
\arguments{
\item{src}{str the name of the file to scrape}

\item{dst}{the destination of the scraped data}
}
\value{
NULL invisibly
}
\description{
Scrape the contents of a .qmd to a .R file
}
