% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_covariates.R
\name{select_covariates}
\alias{select_covariates}
\title{Given a species, month and sampling approach select variables for each month}
\usage{
select_covariates(
  mon = "Jan",
  scientificname = "Mola mola",
  path = data_path(),
  must_have_depth = TRUE
)
}
\arguments{
\item{mon}{chr month abbreviation ("Jan" default)}

\item{scientificname}{chr the species studied (default "Mola mola")}

\item{path}{chr file path to the personal data directory}

\item{must_have_depth}{logical, if TRUE then make sure depth is a covariate}
}
\value{
a list suitable for storing as a YAML configuration
}
\description{
Given a species, month and sampling approach select variables for each month
}
