% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{standard_obis_fields}
\alias{standard_obis_fields}
\title{Retrieve a vector of standard OBIS fields}
\usage{
standard_obis_fields()
}
\value{
character vector of names
}
\description{
Retrieve a vector of standard OBIS fields
}
