% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickman.R
\name{stars_as_months}
\alias{stars_as_months}
\title{Create a "month" stars object modeled after a template}
\usage{
stars_as_months(template, use = c("number", "name")[1], factored = FALSE)
}
\arguments{
\item{template}{a stars object with one or more month bands as a 
3rd dimension}

\item{use}{str, one of "number" or "name" to determine how months are identified}

\item{factored}{logical, if TRUE make the attribute a factor}
}
\value{
a stars object with month values (possibly factored) and a month dimension
}
\description{
Create a "month" stars object modeled after a template
}
