% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{threshold_prediction}
\alias{threshold_prediction}
\title{Threshold a prediction map}
\usage{
threshold_prediction(x, threshold = 0.5)
}
\arguments{
\item{x}{stars, a stars prediction map (habitat suitability index)}

\item{threshold}{numeric, values in x at or above this value are considered
to be predicted presences}
}
\value{
a stars object with habitat suitability index converted to a binary class
}
\description{
Threshold a prediction map
}
