% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{tidy_obis}
\alias{tidy_obis}
\title{Tidy raw OBIS results and return as an sf object}
\usage{
tidy_obis(
  x = fetch_obis(),
  fields = c("id", "basisOfRecord", "eventDate", "eventTime", "individualCount",
    "decimalLongitude", "decimalLatitude"),
  crs = 4326
)
}
\arguments{
\item{x}{data frame or tibble of raw obis results}

\item{fields}{chr, the preferred fields and order of fields}

\item{crs}{the CRS of the data}
}
\value{
sf object
}
\description{
Tidy raw OBIS results and return as an sf object
}
