% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{workflowset_selectomatic}
\alias{workflowset_selectomatic}
\title{Selects best hyperparams, transfers those to the final model and then 
performs a final fit and testing suite for each wflow in the set.}
\usage{
workflowset_selectomatic(
  wflow,
  splitdata,
  filename = "workflowset",
  path = data_path("models")
)
}
\arguments{
\item{wflow}{a workflow set with one or more workflows}

\item{splitdata}{rsplit object with training and testing data}

\item{filename}{str or NULL, a name for the file or NULL to not write}

\item{path}{str or NULL, if not NULL then write each fitted workflow to file}
}
\value{
a tibble with one or more  "last_fit" model objects
}
\description{
Note that this wraps a number of important steps into one.  We provided this
in the interest of speed, but the details are worthy of investogating.
}
