% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{write_ensemble}
\alias{write_ensemble}
\title{Write an ensemble given the ensemble, version and path}
\usage{
write_ensemble(ensemble, version = "g008", path = data_path("models"))
}
\arguments{
\item{ensemble}{a SDM ensemble}

\item{version}{chr the version to write}

\item{path}{chr the path to the data directory}
}
\value{
ensemble itself
}
\description{
Write an ensemble given the ensemble, version and path
}
