% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{write_model}
\alias{write_model}
\title{Write an model given the model, version and path}
\usage{
write_model(model, version = "g_Aug", path = data_path("models"))
}
\arguments{
\item{model}{model object}

\item{version}{chr the version to write}

\item{path}{chr the path to the data directory}
}
\value{
model itself
}
\description{
Write an model given the model, version and path
}
