% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowsets.R
\name{write_model_fit}
\alias{write_model_fit}
\title{Write a workflowset fit object to file}
\usage{
write_model_fit(
  x,
  filename = "model_fits",
  path = data_path("models"),
  packed = FALSE
)
}
\arguments{
\item{x}{workflowset object}

\item{filename}{str a name for the file}

\item{path}{str or NULL, if not NULL then write each fitted workflow to file}

\item{packed}{logical, if TRUE then butcher the models before saving}
}
\value{
the input workflowset
}
\description{
Write a workflowset fit object to file
}
