% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{write_workflow}
\alias{write_workflow}
\title{Write a workflow given the workflow, version and path}
\usage{
write_workflow(
  x,
  version = "Aug",
  type = get_model_type(x),
  path = data_path("models")
)
}
\arguments{
\item{x}{workflow object}

\item{version}{str the version to write}

\item{path}{str the path to the data directory}
}
\value{
workflow itself is returned
}
\description{
Write a workflow given the workflow, version and path
}
