% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_stars}
\alias{predict_stars}
\title{Predict a classification from stars object}
\usage{
predict_stars(
  x,
  newdata,
  wids = dplyr::pull(x, dplyr::all_of("wflow_id")),
  type = "prob",
  threshold = 0.5,
  ...
)
}
\arguments{
\item{x}{a workflow set table}

\item{newdata}{stars data}

\item{wids}{str one or more workflow ids}

\item{threshold}{num the aribitray threshold used to define the outcome}

\item{...}{other arguments for `predict.model_fit`}

\item{typ}{chr the type of prediction, by default "prob" - see `?predict.model_fit`}
}
\value{
a stars object
}
\description{
Predict a classification from stars object
}
