% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_input.R
\name{read_model_input}
\alias{read_model_input}
\title{Reads a model input file given species, month, approach and path}
\usage{
read_model_input(
  scientificname = "Mola mola",
  mon = "Jan",
  approach = "greedy",
  path = data_path("model_input")
)
}
\arguments{
\item{scientificname}{chr, the species name}

\item{mon}{chr month abbreviation}

\item{approach}{chr, one of "greedy" or "conservative"}

\item{path}{chr the path to the data directory}
}
\description{
Reads a model input file given species, month, approach and path
}
