% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{read_prediction}
\alias{read_prediction}
\title{Read a prediction (or a time series of predictions)}
\usage{
read_prediction(filename, attr = ".pred_presence", time = NULL)

read_prediction(filename, attr = ".pred_presence", time = NULL)
}
\arguments{
\item{filename}{chr, the name of the file to write}

\item{attr}{chr the name to apply to the attribute}

\item{time}{any type, an optional "time" to apply the the third dimension for a 
mutil-layer array}

\item{file}{chr, the name of the file to write}
}
\value{
the input object x (suitable for piping)

stars object
}
\description{
Read a prediction (or a time series of predictions)

Read a prediction (or a time series of predictions)
}
