% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{read_workflow}
\alias{read_workflow}
\title{Read a workflow given the version and path}
\usage{
read_workflow(version = "g_Aug", path = data_path("models"))
}
\arguments{
\item{version}{chr the version to read}

\item{path}{chr the path to the data directory}
}
\value{
workflow
}
\description{
Read a workflow given the version and path
}
