% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_covariates.R
\name{select_covariates}
\alias{select_covariates}
\title{Given a species, month and sampling approach select variabkes for each month}
\usage{
select_covariates(
  approach = "greedy",
  mon = "Jan",
  scientificname = "Mola mola",
  path = data_path()
)
}
\arguments{
\item{approach}{chr one of "greedy" (default) or "conservative"}

\item{mon}{chr month abbreviation ("Jan" default)}

\item{scientificname}{chr the species studied (default "Mola mola")}

\item{path}{chr file path to the personal data directory}
}
\value{
a
}
\description{
Given a species, month and sampling approach select variabkes for each month
}
