% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{write_prediction}
\alias{write_prediction}
\title{Save a prediction (or a time series of predictions)}
\usage{
write_prediction(x, filename = "prediction.tif", attr = ".pred_presence")
}
\arguments{
\item{x}{stars object with one or more time layers}

\item{filename}{chr, the name of the file to write}

\item{attr}{chr or num, either the positional index or attribute name to write.}
}
\value{
the input object x (suitable for piping)
}
\description{
Save a prediction (or a time series of predictions)
}
