% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{write_workflow}
\alias{write_workflow}
\title{Write a workflow given the workflow, version and path}
\usage{
write_workflow(x, version = "g_Aug", path = data_path("models"))
}
\arguments{
\item{x}{workflow object}

\item{version}{chr the version to write}

\item{path}{chr the path to the data directory}
}
\value{
workflow itself is returned
}
\description{
Write a workflow given the workflow, version and path
}
